Incorporate blog and simplify site
This change incorporates posts from our separate blog repository and
remove the old landing page. This was done in an effort to simplify the
maintenance of the TEECOMlabs blog, and also give us an opportunity to
start fresh.

The Lobby Transit and Room Phone pages still exist, but they are no
longer referenced from the main landing page. As we revisit the design
and content of the site, we'll likely re-establish these links.